+++ GeneralsMD/Code/GameEngine/Source/Common/RTS/Team.cpp
@@ -330,10 +330,8 @@ Call team->setActive() when all members are added. */
 Team *TeamFactory::createInactiveTeam(const AsciiString& name)
 {
 	TeamPrototype *tp = findTeamPrototype(name);
-	if (!tp) {
-		DEBUG_CRASH(( "Team prototype '%s' does not exist", name.str() ));
-		return NULL;
-	}
+	if (!tp)
+		throw ERROR_BAD_ARG;
 
 	Team *t = NULL;
 	if (tp->getIsSingleton())
@@ -364,11 +362,9 @@ Team *TeamFactory::createInactiveTeam(const AsciiString& name)
 // ------------------------------------------------------------------------
 Team *TeamFactory::createTeam(const AsciiString& name)
 {
-	Team *t = createInactiveTeam(name);
-
-	if (t)
-		t->setActive();
-
+	Team *t = NULL;
+	t = createInactiveTeam(name);
+	t->setActive();
 	return t;
 }
 
@@ -849,13 +845,19 @@ TeamPrototype::~TeamPrototype()
 	if (m_factory)
 		m_factory->removeTeamPrototypeFromList(this);
 
-	deleteInstance(m_productionConditionScript);
+	if (m_productionConditionScript)
+	{
+		deleteInstance(m_productionConditionScript);
+	}
 	m_productionConditionScript = NULL;
 
 	for (Int i = 0; i < MAX_GENERIC_SCRIPTS; ++i)
 	{
-		deleteInstance(m_genericScriptsToRun[i]);
-		m_genericScriptsToRun[i] = NULL;
+		if (m_genericScriptsToRun[i])
+		{
+			deleteInstance(m_genericScriptsToRun[i]);
+			m_genericScriptsToRun[i] = NULL;
+		}
 	}
 }
