+++ GeneralsMD/Code/GameEngine/Source/GameLogic/AI/AIPathfind.cpp
@@ -3951,9 +3951,10 @@ void Pathfinder::reset( void )
 	debugPathPos.y = 0.0f;
 	debugPathPos.z = 0.0f;
 
-	deleteInstance(debugPath);
-	debugPath = NULL;
+	if (debugPath)
+		deleteInstance(debugPath);
 
+	debugPath = NULL;
 	m_frameToShowObstacles = 0;
 
 	for (m_queuePRHead=0; m_queuePRHead<PATHFIND_QUEUE_LEN; m_queuePRHead++) {
@@ -8414,10 +8415,10 @@ Bool Pathfinder::slowDoesPathExist( Object *obj,
 	Path *path = findPath(obj, locoSet, from, to);
 	m_ignoreObstacleID = INVALID_ID;
 	Bool found = (path!=NULL);
-
-	deleteInstance(path);
-	path = NULL;
-
+	if (path) {
+		deleteInstance(path);
+		path = NULL;
+	}
 	return found;
 }
 
@@ -9445,7 +9446,9 @@ void Pathfinder::setDebugPath(Path *newDebugpath)
 	if (TheGlobalData->m_debugAI)
 	{
 		// copy the path for debugging
-		deleteInstance(debugPath);
+		if (debugPath)
+			deleteInstance(debugPath);
+
 		debugPath = newInstance(Path);
 					
 		for( PathNode *copyNode = newDebugpath->getFirstNode(); copyNode; copyNode = copyNode->getNextOptimized() )
